+++ src/backend/collection.rs
@@ -336,6 +336,7 @@ fn is_top_level_item_label(label: &str) -> bool {
             | "Person"
             | "Resource"
             | "Specify"
+            | "Relation"
     )
 }